+++ sync-for-payroll/RELEASES.md
@@ -77,3 +77,13 @@ Based on:
 - [csharp v2.2.0] sync-for-payroll
 ### Releases
 - [NuGet v2.2.0] https://www.nuget.org/packages/Codat.Sync.Payroll/2.2.0 - sync-for-payroll
+
+## 2023-11-27 01:30:12
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payroll.yaml
+- Speakeasy CLI 1.122.0 (2.195.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v3.0.0] sync-for-payroll
+### Releases
+- [NuGet v3.0.0] https://www.nuget.org/packages/Codat.Sync.Payroll/3.0.0 - sync-for-payroll